rename reload-network to restart-network
